Moved SMS implementations into proper folders ~ changed folder structure to have 2 packages: one generic and one specific for an sms network ~ Fixed import errors ~ Improved some documentation for Commands ~ Improved Documentation for SMSNetDictionary key validity
