+++ networklibrary/src/main/java/com/eis/communication/network/CommandExecutor.java
@@ -1,7 +1,5 @@
 package com.eis.communication.network;
 
-import com.eis.communication.network.commands.Command;
-
 /**
  * Class that manage all the sent commands.
  * It just lets them executing.
