+++ networklibrary/src/main/java/com/eis/smsnetwork/SMSNetSubscriberList.java
@@ -1,4 +1,5 @@
-package com.eis.communication.network.smsnetwork;
+package com.eis.smsnetwork;
+
 
 import com.eis.communication.network.NetSubscriberList;
 import com.eis.smslibrary.SMSPeer;
