+++ networklibrary/src/test/java/com/eis/smsnetwork/SMSNetDictionaryTest.java
@@ -1,6 +1,6 @@
 package com.eis.communication.network;
 
-import com.eis.communication.network.smsnetwork.SMSNetDictionary;
+import smsnetwork.SMSNetDictionary;
 
 import org.junit.Before;
 import org.junit.Test;
