+++ networklibrary/src/test/java/com/eis/smsnetwork/commands/AddPeerTest.java
@@ -1,7 +1,8 @@
-package com.eis.communication.network.commands;
+package com.eis.smsnetwork.commands;
 
-import com.eis.communication.network.smsnetwork.SMSNetworkManager;
 import com.eis.smslibrary.SMSPeer;
+import com.eis.smsnetwork.SMSNetworkManager;
+import com.eis.smsnetwork.commands.AddPeer;
 
 import org.junit.Test;
