+++ networklibrary/src/test/java/com/eis/smsnetwork/commands/AddResourceTest.java
@@ -1,6 +1,7 @@
-package com.eis.communication.network.commands;
+package com.eis.smsnetwork.commands;
 
-import com.eis.communication.network.smsnetwork.SMSNetworkManager;
+import com.eis.smsnetwork.SMSNetworkManager;
+
 
 import org.junit.Test;
